Fix VBScript character encoding bug
Internally, VBScript strings are stored in Unicode (UTF-16LE). When reading from an ASCII stream, each character is converted to Unicode, and so takes up two bytes instead of one. Therefore, using the AscB function to obtain an ASCII character's code is incorrect: the function returns the value of the first byte of the string parameter, yet the Unicode representation is two bytes in length. A binary file should still be read as an ASCII stream, because we want to read only a single byte at a time. However, obtaining the byte value should be done using the Asc function, as it will convert the Unicode character back to its ASCII representation, and return its code.

See https://support.microsoft.com/en-us/kb/145745 for more information.